Renamed etherscan to snowtrace to avoid package collisions
