+++ setup.py
@@ -9,11 +9,11 @@ setup(
     author_email="info@ethmint.com",
     license="MIT",
     packages=[
-        "etherscan",
-        "etherscan.configs",
-        "etherscan.enums",
-        "etherscan.modules",
-        "etherscan.utils",
+        "snowtrace",
+        "snowtrace.configs",
+        "snowtrace.enums",
+        "snowtrace.modules",
+        "snowtrace.utils",
     ],
     install_requires=["requests"],
     include_package_data=True,
