+++ snowtrace/__init__.py
@@ -1,4 +1,4 @@
-from .etherscan import Etherscan as Snowtrace
+from .snowtrace import Snowtrace
 from .modules.accounts import Accounts as accounts
 from .modules.blocks import Blocks as blocks
 from .modules.contracts import Contracts as contracts
