+++ snowtrace/modules/stats.py
@@ -1,6 +1,6 @@
-from etherscan.enums.actions_enum import ActionsEnum as actions
-from etherscan.enums.fields_enum import FieldsEnum as fields
-from etherscan.enums.modules_enum import ModulesEnum as modules
+from snowtrace.enums.actions_enum import ActionsEnum as actions
+from snowtrace.enums.fields_enum import FieldsEnum as fields
+from snowtrace.enums.modules_enum import ModulesEnum as modules
 
 
 class Stats:
